/album/detail 데이터베이스 가져오는 로직 완성

+++ src/main/java/my/sunghyuk/lifemusic/service/AlbumService.java
@@ -6,7 +6,7 @@ import my.sunghyuk.lifemusic.domain.Album;
 
 public interface AlbumService {
 	
-	Album getAlbumById(int id);
+	Album getAlbumById(long id);
 	List<Album> getAllAlbums();
 	
 }
